+++ R/add_expression_col.R
@@ -1,5 +1,5 @@
 #' @title Template for expressions with statistical details
-#' @name expr_template
+#' @name add_expression_col
 #'
 #' @description
 #'
@@ -69,7 +69,7 @@
 #'
 #' # expression for *t*-statistic with Cohen's *d* as effect size
 #' # note that the plotmath expressions need to be quoted
-#' expr_template(
+#' add_expression_col(
 #'   data           = stats_df,
 #'   statistic.text = list(quote(italic("t"))),
 #'   effsize.text   = list(quote(italic("d"))),
@@ -81,23 +81,23 @@
 #' @export
 
 # function body
-expr_template <- function(data,
-                          paired = FALSE,
-                          statistic.text = NULL,
-                          effsize.text = NULL,
-                          top.text = NULL,
-                          prior.type = NULL,
-                          n = NULL,
-                          n.text = ifelse(
-                            paired,
-                            list(quote(italic("n")["pairs"])),
-                            list(quote(italic("n")["obs"]))
-                          ),
-                          conf.method = "HDI",
-                          k = 2L,
-                          k.df = 0L,
-                          k.df.error = k.df,
-                          ...) {
+add_expression_col <- function(data,
+                               paired = FALSE,
+                               statistic.text = NULL,
+                               effsize.text = NULL,
+                               top.text = NULL,
+                               prior.type = NULL,
+                               n = NULL,
+                               n.text = ifelse(
+                                 paired,
+                                 list(quote(italic("n")["pairs"])),
+                                 list(quote(italic("n")["obs"]))
+                               ),
+                               conf.method = "HDI",
+                               k = 2L,
+                               k.df = 0L,
+                               k.df.error = k.df,
+                               ...) {
   # is this Bayesian test?
   bayesian <- ifelse("bf10" %in% names(data), TRUE, FALSE)
 
@@ -120,10 +120,10 @@ expr_template <- function(data,
   }
 
   # convert needed columns to character type
-  df <- .data_to_char(data, k, k.df, k.df.error)
+  df_expr <- .data_to_char(data, k, k.df, k.df.error)
 
   # adding a few other columns
-  df %<>% mutate(
+  df_expr %<>% mutate(
     statistic.text = statistic.text %||% stat_text_switch(method),
     es.text = effsize.text %||% estimate_type_switch(effectsize),
     prior.type = prior.type %||% prior_type_switch(method),
@@ -136,13 +136,13 @@ expr_template <- function(data,
 
   if (bayesian) {
     if (is.null(top.text)) {
-      df %<>% mutate(expression = glue("list(
+      df_expr %<>% mutate(expression = glue("list(
             log[e]*(BF['01'])=='{format_value(-log(bf10), k)}',
             {es.text}[{prior.type}]^'posterior'=='{estimate}',
             CI['{conf.level}']^{conf.method}~'['*'{conf.low}', '{conf.high}'*']',
             {prior.distribution}=='{prior.scale}')"))
     } else {
-      df %<>% mutate(expression = glue("list(
+      df_expr %<>% mutate(expression = glue("list(
             atop('{top.text}',
             list(log[e]*(BF['01'])=='{format_value(-log(bf10), k)}',
             {es.text}[{prior.type}]^'posterior'=='{estimate}',
@@ -157,7 +157,7 @@ expr_template <- function(data,
   # 0 degrees of freedom --------------------
 
   if (!bayesian && no.parameters == 0L) {
-    df %<>% mutate(expression = glue("list(
+    df_expr %<>% mutate(expression = glue("list(
             {statistic.text}=='{statistic}', italic(p)=='{p.value}',
             {es.text}=='{estimate}', CI['{conf.level}']~'['*'{conf.low}', '{conf.high}'*']',
             {n.text}=='{n.obs}')"))
@@ -167,9 +167,9 @@ expr_template <- function(data,
 
   if (!bayesian && no.parameters == 1L) {
     # for chi-squared statistic
-    if ("df" %in% names(df)) df %<>% mutate(df.error = df)
+    if ("df" %in% names(df_expr)) df_expr %<>% mutate(df.error = df)
 
-    df %<>% mutate(expression = glue("list(
+    df_expr %<>% mutate(expression = glue("list(
             {statistic.text}*'('*{df.error}*')'=='{statistic}', italic(p)=='{p.value}',
             {es.text}=='{estimate}', CI['{conf.level}']~'['*'{conf.low}', '{conf.high}'*']',
             {n.text}=='{n.obs}')"))
@@ -178,100 +178,34 @@ expr_template <- function(data,
   # 2 degrees of freedom -----------------
 
   if (!bayesian && no.parameters == 2L) {
-    df %<>% mutate(expression = glue("list(
+    df_expr %<>% mutate(expression = glue("list(
             {statistic.text}({df}, {df.error})=='{statistic}', italic(p)=='{p.value}',
             {es.text}=='{estimate}', CI['{conf.level}']~'['*'{conf.low}', '{conf.high}'*']',
             {n.text}=='{n.obs}')"))
   }
 
   # return the formatted expression
-  parse(text = df$expression[[1]])
+  data %>%
+    polish_data(.) %>%
+    mutate(expression = list(parse(text = df_expr$expression[[1]])))
 }
 
 
+# utilities -----------------
+
+#' Helper function to convert certain numeric columns to character type
 #' @noRd
 
-stat_text_switch <- function(x) {
-  # to make life easier
-  x <- tolower(x)
-
-  # need to list because `case_when` can't handle outputs of different lengths
-  case_when(
-    grepl("^one sample|^two sample|^pair|pearson correlation$", x) ~ list(quote(italic("t")["Student"])),
-    grepl("^boot", x) ~ list(quote(italic("t")["bootstrapped"])),
-    grepl("^welch", x) ~ list(quote(italic("t")["Welch"])),
-    grepl("wilcoxon rank", x) ~ list(quote(italic("W")["Mann-Whitney"])),
-    grepl("wilcoxon signed", x) ~ list(quote(italic("V")["Wilcoxon"])),
-    grepl("afex| of means$", x) ~ list(quote(italic("F")["Fisher"])),
-    grepl("variances", x) ~ list(quote(italic("F")["Welch"])),
-    grepl("friedman", x) ~ list(quote(chi["Friedman"]^2)),
-    grepl("kruskal", x) ~ list(quote(chi["Kruskal-Wallis"]^2)),
-    grepl("spearman", x) ~ list(quote(italic("S"))),
-    grepl("yuen", x) ~ list(quote(italic("t")["Yuen"])),
-    grepl("heteroscedastic", x) ~ list(quote(italic("F")["trimmed-means"])),
-    grepl("probabilities", x) ~ list(quote(chi["gof"]^2)),
-    grepl("pearson's chi", x) ~ list(quote(chi["Pearson"]^2)),
-    grepl("mcnemar's chi", x) ~ list(quote(chi["McNemar"]^2)),
-    grepl("meta", x) ~ list(quote(italic("z"))),
-    TRUE ~ list(NULL)
-  )
+.data_to_char <- function(data, k = 2L, k.df = 0L, k.df.error = 0L) {
+  data %>%
+    mutate(
+      across(.fns = ~ format_value(.x, k), .cols = matches("^est|^sta|p.value|.scale$|.low$|.high$|^log")),
+      across(.fns = ~ format_value(.x, k.df), .cols = matches("^df$")),
+      across(.fns = ~ format_value(.x, k.df.error), .cols = matches("^df.error$")),
+      across(.fns = ~ paste0(.x * 100, "%"), .cols = matches("^conf.level$"))
+    )
 }
 
-#' @noRd
-
-estimate_type_switch <- function(x) {
-  # to make life easier
-  x <- tolower(x)
-
-  # need to list because `case_when` can't handle outputs of different lengths
-  case_when(
-    grepl("pearson's c", x) ~ list(quote(widehat(italic("C"))["Pearson"])),
-    grepl("^pearson", x) ~ list(quote(widehat(italic("r"))["Pearson"])),
-    grepl("spearman", x) ~ list(quote(widehat(rho)["Spearman"])),
-    grepl("^winsor", x) ~ list(quote(widehat(italic("r"))["Winsorized"])),
-    grepl("cohen's d", x) ~ list(quote(widehat(italic("d"))["Cohen"])),
-    grepl("hedges' g", x) ~ list(quote(widehat(italic("g"))["Hedges"])),
-    grepl("^eta2", x) ~ list(quote(widehat(eta["p"]^2))),
-    grepl("^omega2", x) ~ list(quote(widehat(omega["p"]^2))),
-    grepl("biserial", x) ~ list(quote(widehat(italic("r"))["biserial"]^"rank")),
-    grepl("^trimmed", x) ~ list(quote(widehat(mu)["trimmed"])),
-    grepl("^kendall", x) ~ list(quote(widehat(italic("W"))["Kendall"])),
-    grepl("^epsilon2", x) ~ list(quote(widehat(epsilon)["ordinal"]^2)),
-    grepl("^cramers_v$", x) ~ list(quote(italic("V"))),
-    grepl("cramer", x) ~ list(quote(widehat(italic("V"))["Cramer"])),
-    grepl("cohen's g", x) ~ list(quote(widehat(italic("g"))["Cohen"])),
-    grepl("^explanatory", x) ~ list(quote(widehat(xi))),
-    grepl("difference$", x) ~ list(quote(widehat(delta)["R"]^"AKP")),
-    grepl("average$", x) ~ list(quote(widehat(delta)["R-avg"]^"AKP")),
-    grepl("^bayesian pearson", x) ~ list(quote(rho)),
-    grepl("posterior|t-", x) ~ list(quote(italic(delta))),
-    grepl("linear", x) ~ list(quote(italic(R^"2"))),
-    grepl("^meta", x) ~ list(quote(widehat(beta)["summary"]^"meta")),
-    TRUE ~ list(NULL)
-  )
-}
-
-#' @noRd
-
-prior_switch <- function(x) {
-  case_when(
-    grepl("contingency", x, TRUE) ~ list(quote(italic("a")["Gunel-Dickey"])),
-    grepl("correlation", x, TRUE) ~ list(quote(italic("r")["beta"]^"JZS")),
-    TRUE ~ list(quote(italic("r")["Cauchy"]^"JZS"))
-  )
-}
-
-#' @noRd
-
-prior_type_switch <- function(x) {
-  case_when(
-    grepl("contingency", x, TRUE) ~ list("Cramer"),
-    grepl("correlation", x, TRUE) ~ list("Pearson"),
-    grepl("t-|meta-", x, TRUE) ~ list("difference"),
-    grepl("linear", x, TRUE) ~ list("Bayesian"),
-    TRUE ~ list(NULL)
-  )
-}
 
 #' @noRd